+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Added graceful degradation for analytics page when Chart.js fails to load.
+
 ## [v18.9.2](https://github.com/lexicalunit/spellbot/releases/tag/v18.9.2) - 2026-05-14
 
 ### Changed